+++ Objects/Plane.cpp
@@ -2,10 +2,10 @@
 
 bool Plane::intersect(const Ray& ray, IntersectionData *intersectionData) const
 {
-    double denom = ray.d.dot(n);
+    float denom = ray.d.dot(n);
     if (denom == 0)
         return false;
-    double d = (p - ray.o).dot(n) / denom;
+    float d = (p - ray.o).dot(n) / denom;
     if (d < DELTA_INTERSECT)
         return false;
