+++ Objects/Sphere.h
@@ -10,14 +10,14 @@ class Sphere : public Object
         EIGEN_MAKE_ALIGNED_OPERATOR_NEW
 
     public:
-        Sphere(Material *mat, double r, const Vector3d& p)
+        Sphere(Material *mat, float r, const Vector3f& p)
         : Object(mat), m_radius(r), m_pos(p) {}
 
         bool intersect(const Ray& ray, IntersectionData *intersectionData) const override;
 
     private:
-        double m_radius;
-        Vector3d m_pos;
+        float m_radius;
+        Vector3f m_pos;
 };
 
 #endif // SPHERE_H
